refactor: subject list component
