+++ src/lib/server.ts
@@ -1,4 +1,5 @@
 import {
+    Courses,
     Tab,
     bcaSemesterList,
     branchList,
@@ -8,24 +9,39 @@ import {
 import axios, { AxiosError, AxiosResponse } from "axios";
 import _ from "lodash";
 
-export const getBtechSubjectList = async ({
+export const getSubjectList = async ({
+    course,
     semester,
     branch,
 }: {
+    course: Courses;
     semester: string | null;
     branch: string | null;
 }) => {
-    if (
-        !semesterList.some((s) => semester === s.label) ||
-        !branchList.some((b) => branch === b.value)
-    ) {
-        throw new AxiosError("Please check semester and branch.");
+    if (course === Courses.BTECH) {
+        if (
+            !semesterList.some((s) => semester === s.label) ||
+            !branchList.some((b) => branch === b.value)
+        ) {
+            throw new AxiosError("Please check semester and branch.");
+        }
+        const response = (await axios.get(
+            `${server}btech/${semesterList.find((s) => semester === s.label)
+                ?.value}/${branchList.find((b) => branch === b.label)?.value}`
+        )) as AxiosResponse;
+        return response.data;
     }
-    const response = (await axios.get(
-        `${server}btech/${semesterList.find((s) => semester === s.label)
-            ?.value}/${branchList.find((b) => branch === b.label)?.value}`
-    )) as AxiosResponse;
-    return response.data;
+
+    if (course === Courses.BCA) {
+        if (!bcaSemesterList.some((s) => semester === s.label)) {
+            throw new AxiosError("Please check semester and branch.");
+        }
+        const response = (await axios.get(
+            `${server}bca/${bcaSemesterList.find((s) => semester === s.label)
+                ?.value}`
+        )) as AxiosResponse;
+        return response.data;
+    }
 };
 
 export const getBtechSubjectDetails = async ({
@@ -100,21 +116,6 @@ export const getBtechStudyMaterial = async ({
     }
 
     return null;
-};
-
-export const getBcaSubjectList = async ({
-    semester,
-}: {
-    semester: string | null;
-}) => {
-    if (!bcaSemesterList.some((s) => semester === s.label)) {
-        throw new AxiosError("Please check semester and branch.");
-    }
-    const response = (await axios.get(
-        `${server}bca/${bcaSemesterList.find((s) => semester === s.label)
-            ?.value}`
-    )) as AxiosResponse;
-    return response.data;
 };
 
 export const getBcaSubjectDetails = async ({
